requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
lxml>=4.9.0
rarfile>=4.0
guessit>=3.7.0
subliminal>=2.1.0
babelfish>=0.6.0
setuptools
